+++ Runtime/Ponito/Core/Paths/PathNameInfo.Ctors.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace google_storage_upload_tool.Models
+namespace Ponito.Core.Paths
 {
     public readonly partial struct PathNameInfo
     {