+++ Runtime/Ponito/Core/Paths/PathNameInfo.Props.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 
-namespace google_storage_upload_tool.Models
+namespace Ponito.Core.Paths
 {
     public readonly partial struct PathNameInfo
     {