+++ Runtime/Ponito/Core/Asyncs/Tasks/Controls.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
@@ -8,9 +8,8 @@ using Ponito.Core.Asyncs.Tasks.Movables;
 
 namespace Ponito.Core.Asyncs.Tasks
 {
-    public partial class PoTask
+    public static class Controls
     {
-        
         [DebuggerHidden]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Movable Yield(CancellationToken ct = default)